Add various MEM, OS, EXI related changes.

+++ std/Include/mem.h
@@ -1,3 +1,3 @@
 #pragma once
 
-void* const memmove(void* dest, const void* source, size_t size);
+void* const memmove(void* dest, const void* source, u64 size);
